Rename editVote to editStake
