+++ src/modules/pos/validator/components/EditStake/editStake.css
@@ -107,6 +107,6 @@
   }
 }
 
-.removeVoteButton {
+.removeStakeButton {
   border-style: solid;
 }
