+++ src/modules/pos/validator/components/EditStake/index.test.js
